+++ src/core/conversion_options.cpp
@@ -7,7 +7,30 @@
 
 #include "conversion_options.h"
 
+#include <iostream>
+
 using namespace d2m;
 
-// Non-specialized class template data member values:
-///template <class T> const ConversionOptionsInfo ConversionOptionsInterface<T>::m_Info = {};
+void ConversionOptionsBase::PrintHelp(ModuleType moduleType)
+{
+    const auto& definitions = Factory<ConversionOptions>::GetInfo(moduleType)->optionDefinitions;
+
+    std::string name = Factory<Module>::GetInfo(moduleType)->fileExtension;
+    if (name.empty())
+        return;
+
+    for (auto& c : name)
+    {
+        c = toupper(c);
+    }
+
+    if (definitions.Count() == 0)
+    {
+        std::cout << name << " files have no conversion options.\n";
+        return;
+    }
+
+    std::cout << name << " Options:\n";
+
+    definitions.PrintHelp();
+}
