--- src/core/module_base.cpp
@@ -1,23 +1,0 @@
-/*
-    module_base.cpp
-    Written by Dalton Messmer <messmer.dalton@gmail.com>.
-
-    See module_base.h.
-*/
-
-#include "module_base.h"
-#include "conversion_options_base.h"
-
-using namespace d2m;
-
-/*
-const std::shared_ptr<const OptionDefinitionCollection>& ModuleBase::GetOptionDefinitions() const
-{
-    return ConversionOptionsBase::GetDefinitions(GetType());
-}
-
-const std::shared_ptr<const OptionDefinitionCollection>& ModuleBase::GetOptionDefinitions(ModuleType moduleType)
-{
-    return ConversionOptionsBase::GetDefinitions(moduleType);
-}
-*/
