+++ src/utils/utils.cpp
@@ -7,7 +7,8 @@
 
 #include "utils.h"
 
-#include "config.h"
+#include "factory.h"
+#include "module.h"
 
 #include <fstream>
 //#include <filesystem>
